+++ docs/README.md
@@ -31,10 +31,10 @@
 - [ ]  구간 추가 api 구현
     - post `/lines/{id}/sections`
     - [x] 상행 종점 id, 하행 종점 id, 거리
-    - 구간을 추가하기 위해서는 역이 등록되어 있어야 함 (하나라도 없으면 에러)
-    - 구간 안에 구간을 삽입할 때 거리 예외 처리
-    - 구간 안에 구간을 삽입 할 때 기존 구간도 변경해야 함
-    - 상행, 하행 모두 등록되어 있을 때 추가할 수 없음
+    - [ ] 구간을 추가하기 위해서는 역이 등록되어 있어야 함 (하나라도 없으면 에러)
+    - [ ] 구간 안에 구간을 삽입할 때 거리 예외 처리
+    - [ ] 구간 안에 구간을 삽입 할 때 기존 구간도 변경해야 함
+    - [ ] 상행, 하행 모두 등록되어 있을 때 추가할 수 없음
 - [x]  노선 조회 시 구간에 포함된 역 목록 응답
     - get `/lines/{id}`
     - [x] stations 정보 추가
