+++ src/main/java/wooteco/subway/service/SectionService.java
@@ -2,20 +2,41 @@ package wooteco.subway.service;
 
 import org.springframework.stereotype.Service;
 import wooteco.subway.domain.section.Section;
+import wooteco.subway.domain.station.Station;
+import wooteco.subway.dto.LineRequest;
+import wooteco.subway.dto.SectionRequest;
 import wooteco.subway.repository.SectionDao;
 
 @Service
 public class SectionService {
 
     private final SectionDao sectionDao;
+    private final StationService stationService;
 
-    public SectionService(SectionDao sectionDao) {
+    public SectionService(SectionDao sectionDao, StationService stationService) {
         this.sectionDao = sectionDao;
+        this.stationService = stationService;
     }
 
-    public Section createSection(Section section, Long lineId) {
-        long id = sectionDao.save(section, lineId);
-        section.setId(id);
+    public Section createSection(LineRequest lineRequest, Long lindId) {
+        Section section = new Section(
+            stationService.findById(lineRequest.getUpStationId()),
+            stationService.findById(lineRequest.getDownStationId()),
+            lineRequest.getDistance()
+        );
+        Long sectionId = sectionDao.save(section, lindId);
+        section.setId(sectionId);
+        return section;
+    }
+
+    public Section createSection(SectionRequest sectionRequest, Long lineId) {
+        Section section = new Section(
+            stationService.findById(sectionRequest.getUpStationId()),
+            stationService.findById(sectionRequest.getDownStationId()),
+            sectionRequest.getDistance()
+        );
+        Long sectionId = sectionDao.save(section, lineId);
+        section.setId(sectionId);
         return section;
     }
 }
