+++ src/test/java/wooteco/subway/domain/section/SectionTest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.section;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
